DRIVE/requirement.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
aiohttp=3.9.5=pypi_0
aiosignal=1.3.1=pypi_0
altair=5.3.0=pypi_0
aom=3.6.0=h6a678d5_0
asttokens=2.4.1=pyhd8ed1ab_0
astunparse=1.6.3=pypi_0
attrs=23.2.0=pypi_0
blas=1.0=mkl
blessed=1.20.0=pypi_0
blosc=1.21.3=h6a678d5_0
brotli=1.0.9=h9c3ff4c_4
brotli-python=1.0.9=py312h6a678d5_8
brunsli=0.1=h9c3ff4c_0
bzip2=1.0.8=h5eee18b_6
c-ares=1.19.1=h5eee18b_0
ca-certificates=2024.7.2=h06a4308_0
cachetools=5.5.0=pypi_0
certifi=2024.8.30=py312h06a4308_0
cfitsio=3.470=hb418390_7
charls=2.2.0=h2531618_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
clip=1.0=pypi_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.2.1=pypi_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.5.39=0
cuda-runtime=12.1.0=0
cuda-version=12.5=3
cycler=0.12.1=pypi_0
dav1d=1.2.1=h5eee18b_0
debugpy=1.6.7=py312h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
executing=2.0.1=pyhd8ed1ab_0
expat=2.6.3=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py312h06a4308_0
flatbuffers=24.3.25=pypi_0
fonttools=4.53.0=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.1=pypi_0
fsspec=2024.6.0=pypi_0
ftfy=6.2.0=pypi_0
gast=0.6.0=pypi_0
giflib=5.2.1=h5eee18b_3
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-pasta=0.2.0=pypi_0
gpustat=1.1.1=pypi_0
gputil=1.4.0=pypi_0
grpcio=1.64.1=pypi_0
h5py=3.11.0=py312h34c39bb_0
hdf5=1.12.1=h2b7332f_3
huggingface-hub=0.23.3=pypi_0
idna=3.7=py312h06a4308_0
imagecodecs=2023.1.23=py312h81b8100_1
imageio=2.34.2=pyh12aca89_0
imageio-ffmpeg=0.5.1=pypi_0
importlib-metadata=7.1.0=pyha770c72_0
importlib_metadata=7.1.0=hd8ed1ab_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.29.4=pyh3099207_0
ipython=8.25.0=pyh707e725_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.4=py312h06a4308_0
joblib=1.4.2=pypi_0
jpeg=9e=h5eee18b_1
jsonschema=4.22.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter_client=8.6.2=pyhd8ed1ab_0
jupyter_core=5.5.0=py312h06a4308_0
jxrlib=1.1=h7b6447c_2
kaleido=0.2.1=pypi_0
keras=3.4.1=pypi_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.5=pypi_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lazy_loader=0.4=pyhd8ed1ab_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.40=hf3520f5_3
lerc=3.0=h295c915_0
libaec=1.0.4=he6710b0_1
libavif=0.11.1=h5eee18b_0
libbrotlicommon=1.0.9=h5eee18b_8
libbrotlidec=1.0.9=h5eee18b_8
libbrotlienc=1.0.9=h5eee18b_8
libclang=18.1.1=pypi_0
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.10.0.4=0
libcurand=10.3.6.39=0
libcurl=8.7.1=h251f7ec_0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libffi=3.4.2=h7f98852_5
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libllvm14=14.0.6=hdb19cb5_3
libnghttp2=1.57.0=h2d74bed_0
libnpp=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libpng=1.6.39=h5eee18b_0
libsodium=1.0.18=h36c2ea0_1
libssh2=1.11.0=h251f7ec_0
libstdcxx-ng=13.2.0=hc0a3c3a_7
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
libzopfli=1.0.3=h9c3ff4c_0
lightning-utilities=0.11.2=pypi_0
llvm-openmp=14.0.6=h9e868ea_0
llvmlite=0.43.0=py312h6a678d5_0
lru-dict=1.3.0=pypi_0
lz4-c=1.9.4=h6a678d5_1
markdown=3.6=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=py312h5eee18b_0
matplotlib=3.9.0=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py312h5eee18b_1
mkl_fft=1.3.8=py312h5eee18b_0
mkl_random=1.2.4=py312hdb19cb5_0
ml-dtypes=0.4.0=pypi_0
mpmath=1.3.0=py312h06a4308_0
multidict=6.0.5=pypi_0
namex=0.0.8=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
networkx=3.2.1=py312h06a4308_0
nltk=3.8.1=pypi_0
numba=0.60.0=py312h526ad5a_0
numpy=1.26.4=py312hc5e2394_0
numpy-base=1.26.4=py312h0da6c21_0
nvidia-ml-py=12.535.161=pypi_0
nvitop=1.3.2=pypi_0
opencv-python=4.10.0.82=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.0.15=h5eee18b_0
opt-einsum=3.3.0=pypi_0
optree=0.12.1=pypi_0
packaging=24.0=pyhd8ed1ab_0
pandas=2.2.2=pypi_0
parso=0.8.4=pyhd8ed1ab_0
patsy=0.5.6=pypi_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.3.0=py312h5eee18b_0
pip=24.0=pyhd8ed1ab_0
platformdirs=4.2.2=pyhd8ed1ab_0
plotly=5.23.0=pypi_0
plotly-express=0.4.1=pypi_0
prompt-toolkit=3.0.46=pyha770c72_0
protobuf=4.25.3=pypi_0
psutil=5.9.0=py312h5eee18b_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pybind11-abi=5=hd3eb1b0_0
pycapnp=2.0.0=pypi_0
pycurl=7.45.3=pypi_0
pygments=2.18.0=pyhd8ed1ab_0
pynvml=11.5.3=pypi_0
pyparsing=3.1.2=pypi_0
pysocks=1.7.1=py312h06a4308_0
python=3.12.3=h996f2a0_1
python-dateutil=2.9.0=pyhd8ed1ab_0
pytorch=2.3.1=py3.12_cuda12.1_cudnn8.9.2_0
pytorch-cuda=12.1=ha16c6d3_5
pytorch-lightning=1.9.4=pypi_0
pytorch-mutex=1.0=cuda
pytz=2024.1=pypi_0
pyyaml=6.0.1=py312h5eee18b_0
pyzmq=25.1.2=py312h6a678d5_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
regex=2024.5.15=pypi_0
requests=2.32.3=pypi_0
rich=13.7.1=pypi_0
rpds-py=0.18.1=pypi_0
safetensors=0.4.3=pypi_0
scikit-image=0.23.2=py312h526ad5a_0
scipy=1.13.1=py312hc5e2394_0
setuptools=70.0.0=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
smbus2=0.4.3=pypi_0
snappy=1.2.1=h6a678d5_0
sqlite=3.45.3=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
statsmodels=0.14.2=pypi_0
sympy=1.12=py312h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tenacity=9.0.0=pypi_0
tensorboard=2.17.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorflow=2.17.0=pypi_0
termcolor=2.4.0=pypi_0
tifffile=2023.2.28=pyhd8ed1ab_0
timm=1.0.3=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.19.1=pypi_0
toolz=0.12.1=pypi_0
torch-tb-profiler=0.4.3=pypi_0
torchaudio=2.3.1=py312_cu121
torchmetrics=1.4.0.post0=pypi_0
torchvision=0.18.1=py312_cu121
tornado=6.3.3=py312h5eee18b_0
tqdm=4.66.4=pypi_0
traitlets=5.14.3=pyhd8ed1ab_0
transformers=4.41.2=pypi_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024.1=pypi_0
urllib3=2.2.1=py312h06a4308_0
wcwidth=0.2.13=pyhd8ed1ab_0
werkzeug=3.0.3=pypi_0
wheel=0.43.0=pyhd8ed1ab_1
wrapt=1.16.0=pypi_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
yarl=1.9.4=pypi_0
zeromq=4.3.5=h6a678d5_0
zfp=1.0.0=h6a678d5_0
zipp=3.18.1=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.5=hc292b87_2
